test(components): added unit tests for newly introduced components
