+++ src/components/TabularReport/DownloadOptions.js
@@ -5,28 +5,30 @@ import styles from '../../utils/styles'
 
 const prefix = i18n.t('Download as')
 
-export const DownloadOption = ({ url, extension }, key) => (
+export const DownloadOption = ({ file }) => (
     <a
         style={styles.downloadButton}
-        href={url}
+        href={file.url}
         download
         tabIndex="0"
-        key={key}
-        target={extension === 'pdf' ? '_blank' : '_self'}
+        target={file.extension === 'pdf' ? '_blank' : '_self'}
     >
-        {prefix} {extension}
+        {prefix} {file.extension}
     </a>
 )
 
 DownloadOption.propTypes = {
-    url: PropTypes.string.isRequired,
-    extension: PropTypes.string.isRequired,
-    key: PropTypes.number.isRequired,
+    file: PropTypes.shape({
+        url: PropTypes.string.isRequired,
+        extension: PropTypes.string.isRequired,
+    }),
 }
 
 export const DownloadOptions = ({ fileUrls }) => (
     <div id="download-options-container" style={styles.downloadContainer}>
-        {fileUrls.map(DownloadOption)}
+        {fileUrls.map((file, index) => (
+            <DownloadOption file={file} key={index} />
+        ))}
     </div>
 )
