+++ src/components/TabularReport/ReportTable.js
@@ -2,11 +2,17 @@
 import React, { Fragment } from 'react'
 import PropTypes from 'prop-types'
 
-const Header = (header, index) => <th key={`h-${header}-${index}`}>{header}</th>
-const Cell = (cell, index) => <td key={`c-${cell}-${index}`}>{cell}</td>
-const Row = (rowItems, index) => (
-    <tr key={`r-${rowItems[0]}-${index}`}>{rowItems.map(Cell)}</tr>
+const Row = ({ cells }) => (
+    <tr>
+        {cells.map((cell, index) => (
+            <td key={`c-${cell}-${index}`}>{cell}</td>
+        ))}
+    </tr>
 )
+
+Row.propTypes = {
+    cells: PropTypes.arrayOf(PropTypes.string).isRequired,
+}
 
 const ReportTable = ({ content: { title, headers, rows } }) => (
     <Fragment>
@@ -14,11 +20,15 @@ const ReportTable = ({ content: { title, headers, rows } }) => (
         <table>
             <thead>
                 <tr className="tabular-report__header-row">
-                    {headers.map(Header)}
+                    {headers.map((header, index) => (
+                        <th key={`h-${header}-${index}`}>{header}</th>
+                    ))}
                 </tr>
             </thead>
             <tbody className={'tabular-report__table-body'}>
-                {rows.map(Row)}
+                {rows.map((cells, index) => (
+                    <Row cells={cells} key={`r-${cells[0]}-${index}`} />
+                ))}
             </tbody>
         </table>
     </Fragment>
