+++ src/redux/selectors/orgUnitDistReport/getTransformedChartData.js
@@ -26,6 +26,7 @@ export default function getTransformedChartData(state) {
     const chartData = getChartData(labels, dataPoints, title)
 
     cache.setCachedResult(data, chartData)
+    console.log(JSON.stringify(chartData, null, 4))
     return chartData
 }
